+++ src/modals/index.js
@@ -6,6 +6,7 @@ var RenameModal = require('./rename-modal');
 var UnsavedSwitchModal = require('./unsaved-switch-modal');
 var HistoryModal = require('./history-modal');
 var HistoryEntryModal = require('./history-entry-modal');
+var TabSwitcherModal = require('./tab-switcher-modal');
 
 module.exports = {
     SessionManagerModal: SessionManagerModal,
@@ -15,4 +16,5 @@ module.exports = {
     HistoryModal: HistoryModal,
     HistoryEntryModal: HistoryEntryModal,
     DeleteOrArchiveModal: require('./delete-or-archive-modal'),
+    TabSwitcherModal: TabSwitcherModal,
 };